Crate dists

Sampling from random distributions.

This is a generalization of Rand to allow parameters to control the exact properties of the generated values, e.g. the mean and standard deviation of a normal distribution. The Sample trait is the most general, and allows for generating values that change some state internally. The IndependentSample trait is for generating values that do not need to record state.

Reexports

pub use self::gamma::{Gamma, ChiSquared, FisherF, StudentT};

pub use self::normal::{Normal, LogNormal};

pub use self::exponential::Exp;

Modules

exponential
The exponential distribution.

gamma
The Gamma and derived distributions.

normal
The normal and derived distributions.

Structs

W

Weighted
A value with a particular weight for use with WeightedChoice.

WeightedChoice
A distribution that selects from a finite collection of weighted items.